+++ PcbReader/src/Layers/Gerber/Entities/Apertures/Macro/Expressions/ValueExpression.cs
@@ -1,8 +1,8 @@
 ﻿namespace PcbReader.Layers.Gerber.Entities.Apertures.Macro.Expressions;
 
 public class ValueExpression: IExpression {
-    public ValueExpression(decimal value) {
+    public ValueExpression(double value) {
         Value = value;
     }
-    public decimal Value { get; set; }
+    public double Value { get; set; }
 }
